Use timeout instead of transitionend for resolving...
...shades' fade-in/out promises.

This may be a temporary thing; but for now I'm at times seeing that the
CSS transition doesn't fire, so then transitionend never fires, and the
promise goes unresolved. I'll look into it at a later date.

This workaround has the downside that if the CSS transition does fail
to fire and the shade instead fades in or out instantly, there will be
a small visual delay while the timeout waits itself out.