Adopt open_modal_dialog() for the existing dialogs
I considered for a moment going for a non-modal approach, where several
changes to an observation (place, date) could be queued up simultaneously
rather than waiting for a modal dialog to finish each time. But I
somewhat prefer the more focused workflow of the modal approach, so for
now it's what I'm going with.