ObservationList: Add a lock symbol to the action bar
The lock symbol conveys to the user whether the list they're viewing is
open or locked for editing.

The open lock symbol will be displayed when viewing a list with the edit
key; and the closed lock symbol when using the view key. In the future,
the open lock will have a link to the view key (public) version of the
list.