ObservationList: Remove obsolete code in...
...generate_observation_elements().

This code was used as a placeholder for testing the application of the
bird search directly to the list of observations such that the search
terms limit what's shown in the list of observations. But that idea was
discarded for now, and the code is fairly unsightly, so for now it's
best removed.