Remove tools/initialize-database-tables.php
This tool became deprecated when the database code was made to use a
class. And I don't really want to expose any functionality in the
database class for creating tables; so I think removing this tool is
the way to go. I'll probably replace it with written instructions on how
to replicate Lintulista's table structure.