ObservationList: Retain observation date's...
...hours, minutes, seconds, milliseconds when setting a new observation
date.

The thing with this is, the dialog for setting the date doesn't allow
one to set the aforementioned elements of the date, only the day, month,
and year.

It used to be that the rest were reset to zero. But when that's done, and
if the list is also being sorted by date, elements in the list may switch
places in an unexpected manner - the user is, after all, not touching the
hours, minutes, etc., but the values of those parameters are nonetheless
changing and they're being used to sort the list.

I might implement a way for the user to adjust the time of the observation
as well and not just the date, but we'll see.