backend-access: Don't refresh local cache of observations...
...by re-fetching from the server when entries are modified. Instead,
update the cache locally based on the modifications sent to the server.
This saves a network request per every modification.

The client-side list may drift out of sync with the server when the
observation data isn't refreshed often enough, but I feel there should
be a separate mechanism for ensuring data freshness, rather than relying
on refreshing the list when the client makes changes. (Another client
might be operating on the list as well, in which case this method of
updating would be insufficient from the get-go.)