QueryObservationDeletion: Prevent dialog redraw on...
...input field change. Instead, will now only redraw when the input bird
name matches or loses its match, rather than on every character typed.

Note that this commit simplifies the dialog also, removing the count of
characters in the input field and the icon embedded in the input field
that indicated when the input was valid. The icon wasn't really needed,
as the 'accept' button is enabled/disabled based on whether the input
is valid; and the count of characters likewise wasn't too useful.

The dialogs are due for a new styling, so it's ok that this commit leaves
the dialog looking a bit messy for now.