Remove button bar from observation cards and integrate...
...it into the search result cards. Namely, the button to remove an
existing observation is now shown in the corner of the result card for
results of birds that are already on the list of observations.

It remains yet to add a button to the result card to set the date of
an observation.

I think 'hiding' these buttons in the result card rather than showing
them on the observation card is better, since it reduces the chance of
accidentally clicking on them while browsing the list; and the actions
the buttons provided - deleting an observation and changing its date -
aren't ones you'd perform too often.