ObservationList: If the list is empty, default to sorting by...
...the 100 Lajia challenge. It looks more interesting to at least have
the challenge cards up rather than a fully empty list.

Note that this commit removes the original intro element; which contained
basic instructions on how to get started with Lintulista. But I think
we can get away with not showing this short tutorial, as the link to the
user guide is fairly discoverable, at the top of the page.