index.html: Regenerate random cards list on window resize
Since the number of cards displayed depends on the window size.

This isn't an ideal solution, though, since it's a bit jarring to have
all of the birds on the list change as the list is regenerated. A better
solution may be to generate a random list of x cards on page load, and
decide on window resize how many of those to display.