Implement a type system
The type system provides a kind of class inheritance interface that
can be used to assert that a given object was created by a given
factory function.

For example, LL_Bird(...) is a factory function that creates a bird
object. The type system then appends some special identifiers to the
object, which can be queried with utility functions like
LL_Bird.is_type(<object>) to verify that the object is of the desired
type.