Implement actions
Actions (LL_Action - name still pending) are intended to minimize
repetitive in-code error checking and to make it easier to produce
condensed but at least somewhat relevant user-facing error messages on
failure.

Each action encapsulates some set of related directives (which together
aim to perform some action, like querying the user for their password
and then logging them in), wrapping these directives in a try/catch and
emitting a pre-set user-directed error message on failure.

This is work in progress.